refactor: move file test to tests folder
